Replace sidebar with top navigation menu
